+++ premake5.lua
@@ -42,6 +42,10 @@ project "Engine"
 		"%{prj.name}/src/Tools",
 		"%{prj.name}/src/UI",
 		"%{prj.name}/src/World",
+		"%{prj.name}/src/Event",
+		"%{prj.name}/src/Scene",
+		"%{prj.name}/src/Sound",
+		"%{prj.name}/src/Input",
 		"%{prj.name}/vendor/spdlog/include",
 		"../SDK/DirectXTK/include",
 		"../SDK/FBXSDK/include",
@@ -131,6 +135,10 @@ project "TestGame"
 		"Engine/src/Tools",
 		"Engine/src/UI",
 		"Engine/src/World",
+		"Engine/src/Event",
+		"Engine/src/Scene",
+		"Engine/src/Sound",
+		"Engine/src/Input",
 
 		"../SDK/DirectXTK/include",
 		"../SDK/FBXSDK/include",
